优化main代码
