--- src/main/global/index.js
@@ -1,10 +1,0 @@
-import axios from 'axios'
-import api from './api'
-
-export default {
-  init(window) {
-    global.api = api
-    global.win = window
-    global.clientApi = axios
-  }
-}
