+++ src/main/index.dev.js
@@ -1,19 +1,6 @@
-/**
- * This file is used specifically and only for development. It installs
- * `electron-debug` & `vue-devtools`. There shouldn't be any need to
- *  modify this file, but it can be used to extend your development
- *  environment.
- */
-
-/* eslint-disable */
-
-// Set environment for development
 process.env.NODE_ENV = 'development'
 
-// Install `electron-debug` with `devtron`
 require('electron-debug')({ showDevTools: false })
-
-// Install `vue-devtools`
 require('electron').app.on('ready', () => {
   let installExtension = require('electron-devtools-installer')
     installExtension.default(installExtension.VUEJS_DEVTOOLS.id)
@@ -23,5 +10,4 @@ require('electron').app.on('ready', () => {
     })
 })
 
-// Require `main` process to boot app
 require('./index')
